renaming demo
