+++ doc/user_guide.md
@@ -18,7 +18,7 @@ The specification of the app under test is provided using the following configur
    
 3. `app_classpath_file`. The name of a text file containing all library dependencies of the app under test. The file should contain a list of jar files (using relative or absolute paths); for example, see the [irs classpath file](../test/data/irs/irsMonoClasspath.txt). 
 
-    If your Java project has a Gradle build file, you can alternatively specify that build file and Tackle-Test will automatically collect the application's library dependencies and create the classpath file. To do this, the `generate` command options `app_build_type`, `app_build_config_file`, and `app_build_settings_file` can be used; for example, see the [demo app build specification](../test/data/demo/tkltest_config.toml). This feature is currently supported for Gradle only; Ant and Maven support will be added in the future.
+    If your Java project has a Gradle build file, you can alternatively specify that build file and Tackle-Test will automatically collect the application's library dependencies and create the classpath file. To do this, the `generate` command options `app_build_type`, `app_build_config_file`, and `app_build_settings_file` can be used; for example, see the [split_join app build specification](../test/data/split_join/tkltest_config.toml). This feature is currently supported for Gradle only; Ant and Maven support will be added in the future.
 
 ### Selecting the test-generation strategy
