--- test/data/demo/app/build.gradle
@@ -1,19 +1,0 @@
-
-plugins {
-    id 'demo.java-application-conventions'
-}
-
-dependencies {
-    implementation 'org.apache.commons:commons-text'
-    implementation project(':utilities')
-}
-
-application {
-    mainClass = 'demo.app.App' // <1>
-}
-
-
-task tkltest_get_dependencies2(type: Copy) {
-    from sourceSets.main.runtimeClasspath
-    into 'jarsdir2'
-}
