--- test/data/demo/app/src/main/java/demo/app/App.java
@@ -1,21 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package demo.app;
-
-import demo.list.LinkedList;
-
-import static demo.utilities.StringUtils.join;
-import static demo.utilities.StringUtils.split;
-import static demo.app.MessageUtils.getMessage;
-
-import org.apache.commons.text.WordUtils;
-
-public class App {
-    public static void main(String[] args) {
-        LinkedList tokens;
-        tokens = split(getMessage());
-        String result = join(tokens);
-        System.out.println(WordUtils.capitalize(result));
-    }
-}
