--- test/data/demo/app/src/main/java/demo/app/MessageUtils.java
@@ -1,10 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package demo.app;
-
-class MessageUtils {
-    public static String getMessage() {
-        return "Hello      World!";
-    }
-}
