--- test/data/demo/buildSrc/build.gradle
@@ -1,8 +1,0 @@
-
-plugins {
-    id 'groovy-gradle-plugin' // <1>
-}
-
-repositories {
-    gradlePluginPortal() // <2>
-}
