--- test/data/demo/buildSrc/src/main/groovy/demo.java-common-conventions.gradle
@@ -1,20 +1,0 @@
-
-plugins {
-    id 'java' // <1>
-}
-
-repositories {
-    mavenCentral() // <2>
-}
-
-dependencies {
-    constraints {
-        implementation 'org.apache.commons:commons-text:1.9' // <3>
-    }
-
-    testImplementation 'org.junit.jupiter:junit-jupiter:5.7.2' // <4>
-}
-
-tasks.named('test') {
-    useJUnitPlatform() // <5>
-}
