--- test/data/demo/buildSrc/src/main/groovy/demo.java-library-conventions.gradle
@@ -1,6 +1,0 @@
-
-plugins {
-    id 'demo.java-common-conventions' // <1>
-
-    id 'java-library' // <2>
-}
