--- test/data/demo/settings.gradle
@@ -1,4 +1,0 @@
-
-rootProject.name = 'demo'
-include('app', 'list', 'utilities')
-
