--- test/data/demo/utilities/build.gradle
@@ -1,12 +1,0 @@
-
-plugins {
-    id 'demo.java-library-conventions'
-}
-
-dependencies {
-    api project(':list')
-}
-task getDeps(type: Copy) {
-    from sourceSets.main.runtimeClasspath
-    into './dependencies/'
-}
