--- test/data/demo/utilities/src/main/java/demo/utilities/JoinUtils.java
@@ -1,20 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package demo.utilities;
-
-import demo.list.LinkedList;
-
-class JoinUtils {
-    public static String join(LinkedList source) {
-        StringBuilder result = new StringBuilder();
-        for (int i = 0; i < source.size(); ++i) {
-            if (result.length() > 0) {
-                result.append(" ");
-            }
-            result.append(source.get(i));
-        }
-
-        return result.toString();
-    }
-}
