+++ test/data/split_join/app/src/main/java/demo/app/App.java
@@ -1,0 +1,19 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package split_join.app;
+
+import split_join.list.LinkedList;
+
+import static split_join.utilities.StringUtils.join;
+import static split_join.utilities.StringUtils.split;
+import static split_join.app.MessageUtils.getMessage;
+
+public class App {
+    public static void main(String[] args) {
+        LinkedList tokens;
+        tokens = split(getMessage());
+        String result = join(tokens);
+        System.out.println(result);
+    }
+}
