+++ test/data/split_join/app/src/main/java/demo/app/MessageUtils.java
@@ -1,0 +1,10 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package split_join.app;
+
+class MessageUtils {
+    public static String getMessage() {
+        return "Hello      World!";
+    }
+}
