+++ test/data/split_join/list/build.gradle
@@ -1,0 +1,9 @@
+
+plugins {
+    id 'java'
+    id 'application'
+
+}
+repositories {
+    mavenCentral()
+}
