+++ test/data/split_join/utilities/src/main/java/demo/utilities/JoinUtils.java
@@ -1,0 +1,20 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package split_join.utilities;
+
+import split_join.list.LinkedList;
+
+class JoinUtils {
+    public static String join(LinkedList source) {
+        StringBuilder result = new StringBuilder();
+        for (int i = 0; i < source.size(); ++i) {
+            if (result.length() > 0) {
+                result.append(" ");
+            }
+            result.append(source.get(i));
+        }
+
+        return result.toString();
+    }
+}
