+++ test/data/split_join/utilities/src/main/java/demo/utilities/StringUtils.java
@@ -1,0 +1,16 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package split_join.utilities;
+
+import split_join.list.LinkedList;
+
+public class StringUtils {
+    public static String join(LinkedList source) {
+        return JoinUtils.join(source);
+    }
+
+    public static LinkedList split(String source) {
+        return SplitUtils.split(source);
+    }
+}
